move to float from decimal. decimal has low performance

+++ TriangleTask/TriangleTask.Tests/TriangleAnalyzerShould.cs
@@ -8,7 +8,7 @@ namespace TriangleTask.Tests
 		[TestCase(10, 1 ,1)]
 		[TestCase(1, 10,1)]
 		[TestCase(1, 1,10)]
-		public void ReturnUndefinedIfOneSideMoreThatSumOfTwo(decimal a, decimal b, decimal c)
+		public void ReturnUndefinedIfOneSideMoreThatSumOfTwo(float a, float b, float c)
 		{
 			var result = TriangleTypeAnalyzer.AnalyzeTriangle(a, b, c);
 			Assert.That(result, Is.EqualTo(TriangleType.Undefined));
@@ -17,7 +17,7 @@ namespace TriangleTask.Tests
 		[TestCase(3, 2,2)]
 		[TestCase(2, 3,2)]
 		[TestCase(2, 2,3)]
-		public void ReturnObtuseIfOneSideSquaredThatSquaredSumOfTwo(decimal a, decimal b, decimal c)
+		public void ReturnObtuseIfOneSideSquaredThatSquaredSumOfTwo(float a, float b, float c)
 		{
 			var result = TriangleTypeAnalyzer.AnalyzeTriangle(a, b, c);
 			Assert.That(result, Is.EqualTo(TriangleType.Obtuse));
